# docs/openapi.macrobrief.v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MacroBrief MVP API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/events": {
      "post": {
        "summary": "Submit KPI event",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["eventId", "eventType", "userId", "occurredAtUtc", "sessionId"],
                "properties": {
                  "eventId": {"type": "string"},
                  "eventType": {"type": "string"},
                  "userId": {"type": "string"},
                  "occurredAtUtc": {"type": "string", "format": "date-time"},
                  "sessionId": {"type": "string"},
                  "symbol": {"type": "string"},
                  "newsEventId": {"type": "string"},
                  "feedback": {"type": "string"},
                  "reasonTag": {"type": "string"},
                  "sourceName": {"type": "string"},
                  "sourceUrl": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Created"},
          "400": {"description": "Bad Request"}
        }
      }
    },
    "/api/v1/internal/events": {
      "get": {
        "summary": "Get recent KPI events (internal)",
        "parameters": [
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 100, "minimum": 1, "maximum": 1000}}
        ],
        "responses": {"200": {"description": "OK"}}
      }
    }
  }
}